Input: add support for the DRV2667 haptic driver
Adding support for the DRV2667 haptic driver.  This device has the ability
to store vibration patterns in RAM and execute them once the GO bit is set.

The initial driver sets a basic waveform in the first waveform sequence and
will play the waveform when the GO bit is set and will continously play the
waveform until the GO bit is unset.

Data sheet is here: http://www.ti.com/product/drv2667

Signed-off-by: Dan Murphy <dmurphy@ti.com>
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>

[ upstream commit: 1c24622572d6ab2ec8e731588d1c131563a64b53 ]