Merge branch 'sunvnet-multi-tx-queue'
Sowmini Varadhan says:

====================
sunvnet: Use multiple Tx queues.

v2: moved tcp fix out of this series per David Miller feedback

The primary objective of this patch-set is to address the suggestion from
  http://marc.info/?l=linux-netdev&m=140790778931563&w=2
With the changes in Patch 2, every vnet_port will get  packets from
a single tx-queue, and flow-control/head-of-line-blocking is
confined to the vnet_ports that share that tx queue (as opposed to
flow-controlling *all* peers).

Patch 1 is an optimization that resets the DATA_READY bit when
we re-enable Rx interrupts.  This optimization lets us exit quickly
from vnet_event_napi() when new data has not triggered an interrupt.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: f0c65567b3c1b23f79e8a49139580a3872a68d1f ]